extract disk_use code to execution_engine_testing/test_support
