+++ execution_engine_testing/test_support/Cargo.toml
@@ -23,6 +23,7 @@ once_cell = "1.8.0"
 rand = "0.8.4"
 rocksdb = { version = "0.18", default-features = false, features = ["zstd"]}
 walkdir = "2"
+tempfile = "3"
 
 [dev-dependencies]
 version-sync = "0.9.3"
